feat: global exception 파일 수정

+++ Backend/src/main/java/com/ssafy/d108/backend/global/exception/CartItemNotFoundException.java
@@ -11,7 +11,7 @@ public class CartItemNotFoundException extends BusinessException {
         super(ErrorCode.CART_ITEM_NOT_FOUND);
     }
 
-    public CartItemNotFoundException(Long cartItemId) {
+    public CartItemNotFoundException(Integer cartItemId) {
         super(ErrorCode.CART_ITEM_NOT_FOUND, "장바구니 항목을 찾을 수 없습니다. ID: " + cartItemId);
     }
 }
